requirements.txt
attrs==21.2.0
azure-functions==1.7.2
certifi==2021.5.30
charset-normalizer==2.0.4
graphviz==0.18
idna==3.2
jmespath==0.10.0
jsonschema==4.1.2
Pillow==8.4.0
pyrsistent==0.18.0
raven==6.10.0
requests==2.26.0
svg.path==4.1
urllib3==1.26.6
anyio==3.6.1; python_version >= "3.7" and python_full_version >= "3.6.2"
asgiref==3.5.1; python_version >= "3.7"
attrs==21.2.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
azure-functions==1.7.2
black==22.3.0; python_full_version >= "3.6.2"
certifi==2021.5.30
charset-normalizer==2.0.4; python_full_version >= "3.5.0"
click==8.1.3; python_version >= "3.7" and python_full_version >= "3.6.2"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.5.0" and platform_system == "Windows"
fastapi==0.78.0; python_full_version >= "3.6.1"
graphviz==0.18; python_version >= "3.6"
h11==0.13.0; python_version >= "3.7"
httptools==0.4.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.2; python_version >= "3.5"
isort==5.10.1; python_full_version >= "3.6.1" and python_version < "4.0"
jmespath==0.10.0; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
jsonschema==4.1.2; python_version >= "3.7"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
pathspec==0.9.0; python_full_version >= "3.6.2"
pillow==8.4.0; python_version >= "3.6"
platformdirs==2.5.2; python_version >= "3.7" and python_full_version >= "3.6.2"
pydantic==1.9.0; python_full_version >= "3.6.1"
pyrsistent==0.18.0; python_version >= "3.6"
python-dotenv==0.20.0; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
raven==6.10.0
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.19.1; python_version >= "3.6" and python_full_version >= "3.6.1"
svg.path==4.1
tomli==2.0.1; python_version < "3.11" and python_full_version >= "3.6.2" and python_version >= "3.7"
typing-extensions==4.2.0; python_version < "3.10" and python_full_version >= "3.6.2" and python_version >= "3.7"
urllib3==1.26.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
uvicorn==0.17.6; python_version >= "3.7"
uvloop==0.16.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.8.2; python_version >= "3.7"
websockets==10.3; python_version >= "3.7"

uvicorn~=0.17.6
fastapi~=0.89.1
pydantic~=1.9.0
jsonpickle~=3.0.1